Split builtin type handling into `types/builtin.rs`
