+++ server/src/types/mod.rs
@@ -1,16 +1,18 @@
 // SPDX-FileCopyrightText: © 2022 ChiselStrike <info@chiselstrike.com>
 
-pub(crate) use self::type_system::{TypeSystem, TypeSystemError};
+pub use self::builtin::BuiltinTypes;
+pub use self::type_system::{TypeSystem, TypeSystemError};
 use crate::datastore::query::truncate_identifier;
 use std::collections::BTreeMap;
 use std::ops::Deref;
 use std::sync::Arc;
 use uuid::Uuid;
 
+mod builtin;
 mod type_system;
 
 #[derive(Clone, Debug, PartialEq)]
-pub(crate) enum Type {
+pub enum Type {
     String,
     Float,
     Boolean,
@@ -19,7 +21,7 @@ pub(crate) enum Type {
 }
 
 impl Type {
-    pub(crate) fn name(&self) -> String {
+    pub fn name(&self) -> String {
         match self {
             Type::Float => "number".to_string(),
             Type::String => "string".to_string(),
@@ -37,7 +39,7 @@ impl From<Entity> for Type {
 }
 
 #[derive(Clone, Debug, PartialEq)]
-pub(crate) enum Entity {
+pub enum Entity {
     /// User defined Custom entity.
     Custom(Arc<ObjectType>),
     /// Built-in Auth entity.
@@ -46,7 +48,7 @@ pub(crate) enum Entity {
 
 impl Entity {
     /// Checks whether `Entity` is Auth builtin type.
-    pub(crate) fn is_auth(&self) -> bool {
+    pub fn is_auth(&self) -> bool {
         matches!(self, Entity::Auth(_))
     }
 }
@@ -87,14 +89,14 @@ impl Deref for Entity {
 /// reload the type system after changes are made to the database.
 ///
 /// This may become a problem if a user has many types, but it is simple, robust, and elegant.
-pub(crate) trait ObjectDescriptor {
+pub trait ObjectDescriptor {
     fn name(&self) -> String;
     fn id(&self) -> Option<i32>;
     fn backing_table(&self) -> String;
     fn api_version(&self) -> String;
 }
 
-pub(crate) struct InternalObject {
+pub struct InternalObject {
     name: &'static str,
     backing_table: &'static str,
 }
@@ -117,7 +119,7 @@ impl ObjectDescriptor for InternalObject {
     }
 }
 
-pub(crate) struct ExistingObject<'a> {
+pub struct ExistingObject<'a> {
     name: String,
     api_version: String,
     backing_table: &'a str,
@@ -125,7 +127,7 @@ pub(crate) struct ExistingObject<'a> {
 }
 
 impl<'a> ExistingObject<'a> {
-    pub(crate) fn new(name: &str, backing_table: &'a str, id: i32) -> anyhow::Result<Self> {
+    pub fn new(name: &str, backing_table: &'a str, id: i32) -> anyhow::Result<Self> {
         let split: Vec<&str> = name.split('.').collect();
 
         anyhow::ensure!(
@@ -163,14 +165,14 @@ impl<'a> ObjectDescriptor for ExistingObject<'a> {
     }
 }
 
-pub(crate) struct NewObject<'a> {
+pub struct NewObject<'a> {
     name: &'a str,
     backing_table: String, // store at object creation time so consecutive calls to backing_table() return the same value
     api_version: &'a str,
 }
 
 impl<'a> NewObject<'a> {
-    pub(crate) fn new(name: &'a str, api_version: &'a str) -> Self {
+    pub fn new(name: &'a str, api_version: &'a str) -> Self {
         let mut buf = Uuid::encode_buffer();
         let uuid = Uuid::new_v4();
         let backing_table = format!("ty_{}_{}", name, uuid.to_simple().encode_upper(&mut buf));
@@ -202,7 +204,7 @@ impl<'a> ObjectDescriptor for NewObject<'a> {
 }
 
 #[derive(Debug, Clone, PartialEq, Eq)]
-pub(crate) enum TypeId {
+pub enum TypeId {
     String,
     Float,
     Boolean,
@@ -212,7 +214,7 @@ pub(crate) enum TypeId {
 }
 
 impl TypeId {
-    pub(crate) fn name(&self) -> String {
+    pub fn name(&self) -> String {
         match self {
             TypeId::Id | TypeId::String => "string".to_string(),
             TypeId::Float => "number".to_string(),
@@ -251,9 +253,9 @@ where
 }
 
 #[derive(Debug)]
-pub(crate) struct ObjectType {
+pub struct ObjectType {
     /// id of this object in the meta-database. Will be None for objects that are not persisted yet
-    pub(crate) meta_id: Option<i32>,
+    pub meta_id: Option<i32>,
     /// Name of this type.
     name: String,
     /// Fields of this type.
@@ -265,11 +267,11 @@ pub(crate) struct ObjectType {
     /// Name of the backing table for this type.
     backing_table: String,
 
-    pub(crate) api_version: String,
+    pub api_version: String,
 }
 
 impl ObjectType {
-    pub(crate) fn new<D: ObjectDescriptor>(
+    pub fn new<D: ObjectDescriptor>(
         desc: D,
         fields: Vec<Field>,
         indexes: Vec<DbIndex>,
@@ -321,31 +323,31 @@ impl ObjectType {
         })
     }
 
-    pub(crate) fn user_fields(&self) -> impl Iterator<Item = &Field> {
+    pub fn user_fields(&self) -> impl Iterator<Item = &Field> {
         self.fields.iter()
     }
 
-    pub(crate) fn all_fields(&self) -> impl Iterator<Item = &Field> {
+    pub fn all_fields(&self) -> impl Iterator<Item = &Field> {
         std::iter::once(&self.chisel_id).chain(self.fields.iter())
     }
 
-    pub(crate) fn has_field(&self, field_name: &str) -> bool {
+    pub fn has_field(&self, field_name: &str) -> bool {
         self.all_fields().any(|f| f.name == field_name)
     }
 
-    pub(crate) fn get_field(&self, field_name: &str) -> Option<&Field> {
+    pub fn get_field(&self, field_name: &str) -> Option<&Field> {
         self.all_fields().find(|f| f.name == field_name)
     }
 
-    pub(crate) fn backing_table(&self) -> &str {
+    pub fn backing_table(&self) -> &str {
         &self.backing_table
     }
 
-    pub(crate) fn name(&self) -> &str {
+    pub fn name(&self) -> &str {
         &self.name
     }
 
-    pub(crate) fn persisted_name(&self) -> String {
+    pub fn persisted_name(&self) -> String {
         format!("{}.{}", self.api_version, self.name)
     }
 
@@ -355,7 +357,7 @@ impl ObjectType {
         to_map.check_populate_from(&source_map)
     }
 
-    pub(crate) fn indexes(&self) -> &Vec<DbIndex> {
+    pub fn indexes(&self) -> &Vec<DbIndex> {
         &self.indexes
     }
 }
@@ -367,16 +369,16 @@ impl PartialEq for ObjectType {
 }
 
 #[derive(Debug, Clone, PartialEq)]
-pub(crate) struct DbIndex {
+pub struct DbIndex {
     /// Id of this index in the meta database. Before it's creation, it will be None.
-    pub(crate) meta_id: Option<i32>,
+    pub meta_id: Option<i32>,
     /// Name of the index in database. Before it's creation, it will be None.
     backing_table: Option<String>,
-    pub(crate) fields: Vec<String>,
+    pub fields: Vec<String>,
 }
 
 impl DbIndex {
-    pub(crate) fn new(meta_id: i32, backing_table: String, fields: Vec<String>) -> Self {
+    pub fn new(meta_id: i32, backing_table: String, fields: Vec<String>) -> Self {
         Self {
             meta_id: Some(meta_id),
             backing_table: Some(backing_table),
@@ -384,7 +386,7 @@ impl DbIndex {
         }
     }
 
-    pub(crate) fn new_from_fields(fields: Vec<String>) -> Self {
+    pub fn new_from_fields(fields: Vec<String>) -> Self {
         Self {
             meta_id: None,
             backing_table: None,
@@ -392,7 +394,7 @@ impl DbIndex {
         }
     }
 
-    pub(crate) fn name(&self) -> Option<String> {
+    pub fn name(&self) -> Option<String> {
         self.meta_id.map(|id| {
             let name = format!(
                 "index_{id}_{}__{}",
@@ -451,14 +453,14 @@ impl<'a> FieldMap<'a> {
 ///
 /// See the [`ObjectDescriptor`] trait for details.
 /// Situations where a new versus existing field are created are similar.
-pub(crate) trait FieldDescriptor {
+pub trait FieldDescriptor {
     fn name(&self) -> String;
     fn id(&self) -> Option<i32>;
     fn ty(&self) -> Type;
     fn api_version(&self) -> String;
 }
 
-pub(crate) struct ExistingField {
+pub struct ExistingField {
     name: String,
     ty_: Type,
     id: i32,
@@ -466,7 +468,7 @@ pub(crate) struct ExistingField {
 }
 
 impl ExistingField {
-    pub(crate) fn new(name: &str, ty_: Type, id: i32, version: &str) -> Self {
+    pub fn new(name: &str, ty_: Type, id: i32, version: &str) -> Self {
         Self {
             name: name.to_owned(),
             ty_,
@@ -494,14 +496,14 @@ impl FieldDescriptor for ExistingField {
     }
 }
 
-pub(crate) struct NewField<'a> {
+pub struct NewField<'a> {
     name: &'a str,
     ty_: Type,
     version: &'a str,
 }
 
 impl<'a> NewField<'a> {
-    pub(crate) fn new(name: &'a str, ty_: Type, version: &'a str) -> anyhow::Result<Self> {
+    pub fn new(name: &'a str, ty_: Type, version: &'a str) -> anyhow::Result<Self> {
         Ok(Self { name, ty_, version })
     }
 }
@@ -525,13 +527,13 @@ impl<'a> FieldDescriptor for NewField<'a> {
 }
 
 #[derive(Clone, Debug, PartialEq)]
-pub(crate) struct Field {
-    pub(crate) id: Option<i32>,
-    pub(crate) name: String,
-    pub(crate) type_id: TypeId,
-    pub(crate) labels: Vec<String>,
-    pub(crate) is_optional: bool,
-    pub(crate) is_unique: bool,
+pub struct Field {
+    pub id: Option<i32>,
+    pub name: String,
+    pub type_id: TypeId,
+    pub labels: Vec<String>,
+    pub is_optional: bool,
+    pub is_unique: bool,
     // We want to keep the default the user gave us so we can
     // return it in `chisel describe`. That's the default that is
     // valid in typescriptland.
@@ -546,7 +548,7 @@ pub(crate) struct Field {
 }
 
 impl Field {
-    pub(crate) fn new<D: FieldDescriptor>(
+    pub fn new<D: FieldDescriptor>(
         desc: D,
         labels: Vec<String>,
         default: Option<String>,
@@ -575,22 +577,22 @@ impl Field {
         }
     }
 
-    pub(crate) fn user_provided_default(&self) -> &Option<String> {
+    pub fn user_provided_default(&self) -> &Option<String> {
         &self.default
     }
 
-    pub(crate) fn default_value(&self) -> &Option<String> {
+    pub fn default_value(&self) -> &Option<String> {
         &self.effective_default
     }
 
-    pub(crate) fn generate_value(&self) -> Option<String> {
+    pub fn generate_value(&self) -> Option<String> {
         match self.type_id {
             TypeId::Id => Some(Uuid::new_v4().to_string()),
             _ => self.default.clone(),
         }
     }
 
-    pub(crate) fn persisted_name(&self, parent_type_name: &ObjectType) -> String {
+    pub fn persisted_name(&self, parent_type_name: &ObjectType) -> String {
         format!(
             "{}.{}.{}",
             self.api_version,
@@ -601,25 +603,25 @@ impl Field {
 }
 
 #[derive(Clone, Debug, PartialEq)]
-pub(crate) struct FieldAttrDelta {
-    pub(crate) type_id: TypeId,
-    pub(crate) default: Option<String>,
-    pub(crate) is_optional: bool,
-    pub(crate) is_unique: bool,
+pub struct FieldAttrDelta {
+    pub type_id: TypeId,
+    pub default: Option<String>,
+    pub is_optional: bool,
+    pub is_unique: bool,
 }
 
 #[derive(Clone, Debug, PartialEq)]
-pub(crate) struct FieldDelta {
-    pub(crate) id: i32,
-    pub(crate) attrs: Option<FieldAttrDelta>,
-    pub(crate) labels: Option<Vec<String>>,
+pub struct FieldDelta {
+    pub id: i32,
+    pub attrs: Option<FieldAttrDelta>,
+    pub labels: Option<Vec<String>>,
 }
 
 #[derive(Clone, Debug, PartialEq)]
-pub(crate) struct ObjectDelta {
-    pub(crate) added_fields: Vec<Field>,
-    pub(crate) removed_fields: Vec<Field>,
-    pub(crate) updated_fields: Vec<FieldDelta>,
-    pub(crate) added_indexes: Vec<DbIndex>,
-    pub(crate) removed_indexes: Vec<DbIndex>,
+pub struct ObjectDelta {
+    pub added_fields: Vec<Field>,
+    pub removed_fields: Vec<Field>,
+    pub updated_fields: Vec<FieldDelta>,
+    pub added_indexes: Vec<DbIndex>,
+    pub removed_indexes: Vec<DbIndex>,
 }
